+++ auditor/src/routes/advanced_record_filters.rs
@@ -5,7 +5,7 @@
 // http://opensource.org/licenses/MIT>, at your option. This file may not be
 // copied, modified, or distributed except according to those terms.
 
-use crate::domain::{Component, Record, RecordDatabase, ValidAmount, ValidName};
+use crate::domain::{Record, RecordDatabase, ValidAmount, ValidName};
 use chrono::{DateTime, Utc};
 use core::fmt::Debug;
 use sqlx::{PgPool, QueryBuilder, Row};
@@ -88,44 +88,16 @@ pub async fn advanced_record_filtering(
     filters: Filters,
     pool: &PgPool,
 ) -> Result<Vec<Record>, anyhow::Error> {
-    let mut query = QueryBuilder::new("SELECT a.record_id,
-                          m.meta,
-                          css.components,
-                          a.start_time,
-                          a.stop_time,
-                          a.runtime
-                   FROM accounting a
-                   LEFT JOIN (
-                       WITH subquery AS (
-                           SELECT m.record_id as record_id, m.key as key, array_agg(m.value) as values
-                           FROM meta as m
-                           GROUP BY m.record_id, m.key
-                       )
-                       SELECT s.record_id as record_id, array_agg(row(s.key, s.values)) as meta
-                       FROM subquery as s
-                       GROUP BY s.record_id
-                       ) m ON m.record_id = a.record_id
-                   LEFT JOIN (
-                       WITH subquery AS (
-                          SELECT 
-                              c.id as cid,
-                              COALESCE(array_agg(row(s.name, s.value)::score) FILTER (WHERE s.name IS NOT NULL AND s.value IS NOT NULL), '{}'::score[]) as scores
-                          FROM components as c
-                          LEFT JOIN components_scores as cs
-                          ON c.id = cs.component_id
-                          LEFT JOIN scores as s
-                          ON cs.score_id = s.id
-                          GROUP BY c.id
-                       )
-                       SELECT rc.record_id as id, array_agg(row(c.name, c.amount, sq.scores)::component) as components
-                       FROM records_components AS rc
-                       LEFT JOIN components as c
-                       ON rc.component_id = c.id
-                       LEFT JOIN subquery AS sq
-                       ON sq.cid = rc.component_id
-                       GROUP BY rc.record_id
-                   ) css ON css.id = a.id
-               ");
+    let mut query = QueryBuilder::new(
+        "SELECT record_id,
+                  meta,
+                  components,
+                  start_time,
+                  stop_time,
+                  runtime
+           FROM auditor
+               ",
+    );
 
     if filters.start_time.is_some()
         || filters.stop_time.is_some()
@@ -137,7 +109,7 @@ pub async fn advanced_record_filtering(
         query.push(" WHERE ".to_string());
         if let Some(record_id) = &filters.record_id {
             // query string -> a.record_id = '{}' and
-            query.push(" a.record_id = ".to_string());
+            query.push(" record_id = ".to_string());
             query.push_bind(record_id);
             query.push(" and ".to_string());
         }
@@ -146,7 +118,7 @@ pub async fn advanced_record_filtering(
             if let Some(operators) = get_operator(start_time_filters) {
                 for operator in operators {
                     // query string -> a.start_time {} '{}' and
-                    query.push(format!(" a.start_time {} ", operator.0));
+                    query.push(format!(" start_time {} ", operator.0));
                     query.push_bind(operator.1);
                     query.push(" and ".to_string());
                 }
@@ -157,7 +129,7 @@ pub async fn advanced_record_filtering(
             if let Some(operators) = get_operator(stop_time_filters) {
                 for operator in operators {
                     // query string -> a.stop_time {} '{}' and
-                    query.push(format!(" a.stop_time {} ", operator.0));
+                    query.push(format!(" stop_time {} ", operator.0));
                     query.push_bind(operator.1);
                     query.push(" and ".to_string());
                 }
@@ -167,21 +139,24 @@ pub async fn advanced_record_filtering(
         if let Some(meta_filters) = &filters.meta {
             for (key, meta_operator) in meta_filters {
                 if let Some(c) = &meta_operator.c {
-                    // query string -> Array['{}'] = ANY(SELECT r.values FROM unnest(m.meta) AS r(key text, values text[]) WHERE r.key = '{}') and
-                    query.push(" Array[".to_string());
-                    query.push_bind(c);
-                    query.push("] = ANY(SELECT r.values FROM unnest(m.meta) AS r(key text, values text[]) WHERE r.key = ".to_string());
+                    // query string -> meta -> "site_id" @> jsonb_build_array("site1") and
+
+                    query.push(" meta ->  ".to_string());
                     query.push_bind(key);
-                    query.push(" ) ".to_string());
-                    query.push(" and ".to_string());
+                    query.push(" @> jsonb_build_array(".to_string());
+                    query.push_bind(c);
+                    query.push(") ");
+                    query.push(" and ");
                 }
                 if let Some(dnc) = &meta_operator.dnc {
-                    // query string -> (NOT EXISTS (SELECT r.values FROM unnest(m.meta) AS r(key text, values text[]) WHERE r.key = '{}' AND Array['{}'] @> r.values)) and
-                    query.push(" (NOT EXISTS (SELECT r.values FROM unnest(m.meta) AS r(key text, values text[]) WHERE r.key = ".to_string());
+                    // query string -> NOT (meta -> "site_id" @> jsonb_build_array("site_1")) and
+
+                    query.push(" NOT (meta ->  ".to_string());
                     query.push_bind(key);
-                    query.push(" AND Array[".to_string());
+                    query.push(" @> jsonb_build_array(".to_string());
                     query.push_bind(dnc);
-                    query.push("] @> r.values)) and ");
+                    query.push(") ) ");
+                    query.push(" and ");
                 }
             }
         }
@@ -190,15 +165,18 @@ pub async fn advanced_record_filtering(
             for (key, component_operator) in component_filters {
                 if let Some(operators) = get_operator(component_operator) {
                     for operator in operators {
-                        // query string -> EXISTS ( SELECT * FROM unnest(components) AS r WHERE r.name = '{}' and r.amount {} '{}' ) and
-                        query.push(
-                            " EXISTS ( SELECT * FROM unnest(components) AS r WHERE r.name = "
-                                .to_string(),
-                        );
+                        // query string -> components->0->>'name' = "CPU" AND
+                        // (components->0->>'amount')::int >10  and
+
+                        query.push("components->0->>'name' = ");
                         query.push_bind(key);
-                        query.push(format!(" and r.amount {} ", &operator.0));
+                        query.push(format!(
+                            " AND (components->0->>'amount')::int {} ",
+                            &operator.0
+                        ));
                         query.push_bind(operator.1);
-                        query.push(" ) and ".to_string());
+
+                        query.push(" and ".to_string());
                     }
                 }
             }
@@ -212,25 +190,25 @@ pub async fn advanced_record_filtering(
             if let Some(operators) = get_operator(runtime_filters) {
                 for operator in operators {
                     // query string ->  a.runtime {} {} and
-                    query.push(format!(" a.runtime {} ", operator.0));
+                    query.push(format!(" runtime {} ", operator.0));
                     query.push_bind(operator.1);
                     query.push(" and ".to_string());
                 }
             }
         } else {
-            query.push(" a.runtime IS NOT NULL".to_string());
+            query.push(" runtime IS NOT NULL".to_string());
         }
     }
 
     if let Some(sort_by) = &filters.sort_by {
         if let SortOption::ASC(asc) = sort_by {
-            query.push(format!(" ORDER BY a.{} ASC", &asc.to_string()));
+            query.push(format!(" ORDER BY {} ASC", &asc.to_string()));
         }
         if let SortOption::DESC(desc) = sort_by {
-            query.push(format!(" ORDER BY a.{} DESC", &desc.to_string()));
+            query.push(format!(" ORDER BY {} DESC", &desc.to_string()));
         }
     } else {
-        query.push(" ORDER BY a.stop_time ".to_string());
+        query.push(" ORDER BY stop_time ".to_string());
     }
 
     if let Some(limit) = &filters.limit {
@@ -285,22 +263,25 @@ pub async fn advanced_record_filtering(
         .await
         .map_err(GetRecordError)?;
 
-    let result: Vec<RecordDatabase> = rows
+    let result: Vec<Record> = rows
         .iter()
-        .map(|row| RecordDatabase {
+        .map(|row| Record {
             record_id: row.try_get("record_id").unwrap(),
-            meta: row.try_get("meta").ok().unwrap_or(None),
-            components: row.try_get("components").ok().unwrap_or(None),
+            meta: row
+                .try_get("meta")
+                .ok()
+                .and_then(|value| serde_json::from_value(value).ok()),
+            components: row
+                .try_get("components")
+                .ok()
+                .and_then(|value| serde_json::from_value(value).ok()),
             start_time: row.try_get("start_time").ok().unwrap_or(None),
             stop_time: row.try_get("stop_time").ok().unwrap_or(None),
             runtime: row.try_get("runtime").ok().unwrap_or(None),
         })
         .collect();
 
-    result
-        .into_iter()
-        .map(Record::try_from)
-        .collect::<Result<Vec<Record>, anyhow::Error>>()
+    Ok(result)
 }
 
 #[tracing::instrument(name = "Getting one record using record_id", skip(record_id, pool))]
@@ -311,53 +292,23 @@ pub async fn get_one_record(
     let is_valid_record_id = ValidName::parse(record_id.clone().to_string());
     if is_valid_record_id.is_ok() {
         Ok(sqlx::query_as!(
-                RecordDatabase,
-                r#"SELECT a.record_id,
-                          m.meta as "meta: Vec<(String, Vec<String>)>",
-                          css.components as "components: Vec<Component>",
-                          a.start_time as "start_time?",
-                          a.stop_time,
-                          a.runtime
-                   FROM accounting a
-                   LEFT JOIN (
-                       WITH subquery AS (
-                           SELECT m.record_id as record_id, m.key as key, array_agg(m.value) as values
-                           FROM meta as m
-                           GROUP BY m.record_id, m.key
-                       )
-                       SELECT s.record_id as record_id, array_agg(row(s.key, s.values)) as meta
-                       FROM subquery as s
-                       GROUP BY s.record_id
-                       ) m ON m.record_id = a.record_id
-                   LEFT JOIN (
-                       WITH subquery AS (
-                          SELECT 
-                              c.id as cid,
-                              COALESCE(array_agg(row(s.name, s.value)::score) FILTER (WHERE s.name IS NOT NULL AND s.value IS NOT NULL), '{}'::score[]) as scores
-                          FROM components as c
-                          LEFT JOIN components_scores as cs
-                          ON c.id = cs.component_id
-                          LEFT JOIN scores as s
-                          ON cs.score_id = s.id
-                          GROUP BY c.id
-                       )
-                       SELECT rc.record_id as id, array_agg(row(c.name, c.amount, sq.scores)::component) as components
-                       FROM records_components AS rc
-                       LEFT JOIN components as c
-                       ON rc.component_id = c.id
-                       LEFT JOIN subquery AS sq
-                       ON sq.cid = rc.component_id
-                       GROUP BY rc.record_id
-                   ) css ON css.id = a.id
-                   WHERE a.record_id = $1
-               "#,
-               &record_id,
-            )
-            .fetch_one(pool)
-            .await
-            .map(Record::try_from)
-            .map_err(GetRecordError)?
-            .ok())
+            RecordDatabase,
+            r#"SELECT record_id,
+                  meta,
+                  components,
+                  start_time,
+                  stop_time,
+                  runtime
+           FROM auditor
+           WHERE record_id = $1
+        "#,
+            &record_id,
+        )
+        .fetch_one(pool)
+        .await
+        .map(Record::try_from)
+        .map_err(GetRecordError)?
+        .ok())
     } else {
         return Ok(None);
     }